refactor: move nostra and zklend states and events to shared. Remove duplicated helpers
