+++ apps/shared/loan_entity/__init__.py
@@ -1,0 +1,2 @@
+from .loan_entity import LoanEntity
+from .nostra import *
