+++ apps/shared/loan_entity/loan_entity.py
