+++ apps/shared/state/__init__.py
@@ -1,0 +1,1 @@
+from .state import State
