+++ apps/shared/state/state.py
@@ -47,7 +47,9 @@ class State(ABC):
     def get_protocol_name(self) -> str:
         """Returns the protocol name for the state"""
         if not self.PROTOCOL_NAME:
-            raise NotImplementedError("PROTOCOL_NAME must be set in the implementing class")
+            raise NotImplementedError(
+                "PROTOCOL_NAME must be set in the implementing class"
+            )
         return self.PROTOCOL_NAME
 
     def process_event(self, method_name: str, event: pd.Series) -> None:
